continuando con Seeders

+++ database/migrations/2015_07_01_185229_create_table_indicadores_metas_efectores_priorizados.php
@@ -16,7 +16,7 @@ class CreateTableIndicadoresMetasEfectoresPriorizados extends Migration {
 			$table->increments('id');
 			$table->string('efector', 14);
 			$table->string('indicador', 5);
-			$table->integer('base')->nullable();
+			$table->float('base')->nullable();
 			$table->float('c1')->nullable();
 			$table->float('c2')->nullable();
 			$table->float('c3')->nullable();
